images-generation/requirements.txt
pydantic>=2
pyyaml
python-dotenv
tenacity
httpx
tqdm
Pillow
openai

